monitor_message/requirements.txt
DrissionPage
pymysql
cachetools
pyahocorasick
rapidfuzz
PyQt5